requirements.txt
simpleaudio
Pillow
pystray
TTS
